add markerUI
